Shimmer animation

+++ app/src/main/java/com/gmail/uli153/akihabara3/data/repositories/BggService.kt
@@ -15,7 +15,7 @@ interface BggService {
     ): BggSearchResponse
 
     @GET("thing")
-    suspend fun getItem(@Query("id") id: Int, @Query("type") type: String? = null): BggItemResponse
+    suspend fun getItem(@Query("id") id: Long, @Query("type") type: String? = null): BggItemResponse
 
     @GET("thing")
     suspend fun getItems(
